fix: 0-23값을 가지는 top_hour 필드의 데이터 타입을 INT -> TINYINT로 변경

+++ src/main/resources/db/migration/V20__create_coupon_usage_stats_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE coupon_usage_stats
     id         BIGINT AUTO_INCREMENT PRIMARY KEY,
     member_id  BIGINT       NOT NULL COMMENT '손님 ID',
     top_dong   VARCHAR(255) NOT NULL COMMENT '쿠폰 사용 상위 동 정보',
-    top_hour   INT          NOT NULL COMMENT '쿠폰 사용 상위 시간대(0~23)',
+    top_hour   TINYINT      NOT NULL COMMENT '쿠폰 사용 상위 시간대(0~23)',
     created_at DATETIME     NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '등록일',
     updated_at DATETIME     NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT '수정일',
